+++ src/main/java/project/como/domain/post/repository/HeartRepository.java
@@ -1,7 +1,7 @@
-package project.como.domain.heart.repository;
+package project.como.domain.post.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import project.como.domain.heart.model.Heart;
+import project.como.domain.post.model.Heart;
 import project.como.domain.post.model.Post;
 import project.como.domain.user.model.User;
