+++ src/main/java/project/como/domain/post/service/PostService.java
@@ -5,9 +5,12 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import project.como.domain.post.exception.HeartConflictException;
+import project.como.domain.post.exception.HeartNotFoundException;
+import project.como.domain.post.model.Heart;
+import project.como.domain.post.repository.HeartRepository;
 import project.como.domain.post.dto.PostCreateRequestDto;
 import project.como.domain.post.dto.PostDetailResponseDto;
 import project.como.domain.post.dto.PostModifyRequestDto;
@@ -18,6 +21,7 @@ import project.como.domain.post.model.Category;
 import project.como.domain.post.model.Post;
 import project.como.domain.post.model.PostState;
 import project.como.domain.post.repository.PostRepository;
+import project.como.domain.user.exception.UserNotFoundException;
 import project.como.domain.user.model.User;
 import project.como.domain.user.repository.UserRepository;
 
@@ -30,6 +34,7 @@ public class PostService {
 	private final int TOTAL_ITEMS_PER_PAGE = 20;
 	private final UserRepository userRepository;
 	private final PostRepository postRepository;
+	private final HeartRepository heartRepository;
 
 	public void createPost(String username, PostCreateRequestDto dto) {
 		User user = userRepository.findByUsername(username).orElseThrow(() -> new IllegalArgumentException("존재하지 않는 유저입니다."));
@@ -108,5 +113,29 @@ public class PostService {
 						.techs(post.getTechs())
 						.build()).toList())
 				.build();
+	}
+
+	public void makeHeart(String username, Long postId) {
+		User user = userRepository.findByUsername(username).orElseThrow(UserNotFoundException::new);
+		Post post = postRepository.findById(postId).orElseThrow(() -> new PostNotFoundException(postId));
+
+		if (heartRepository.findByPostAndUser(post, user).isPresent())
+			throw new HeartConflictException();
+
+		Heart heart = Heart.builder()
+				.user(user)
+				.post(post)
+				.build();
+
+		heartRepository.save(heart);
+	}
+
+	public void deleteHeart(String username, Long postId) {
+		User user = userRepository.findByUsername(username).orElseThrow(UserNotFoundException::new);
+		Post post = postRepository.findById(postId).orElseThrow(() -> new PostNotFoundException(postId));
+
+		Heart heart = heartRepository.findByPostAndUser(post, user).orElseThrow(HeartNotFoundException::new);
+
+		heartRepository.delete(heart);
 	}
 }
